+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add CC lists to Jira trackers and to Bugzilla trackers (OSIDB-2191)
 - Enable flaw draft creation in BZ (OSIDB-2261)
 - Add support for UAT (OSIDB-2447)
+- Added API for Alerts (OSIDB-325)
 
 ### Changed
 - Make workflows API RESTful (OSIDB-1716)